Rename mindtrace.services.base to mindtrace.services.core
